Introduce RayTracingFeatureProcessorInterface to remove dependency from Atom_Feature_Common.Public to .Private.Object
