+++ Gems/Atom/Feature/Common/Code/Source/Mesh/MeshFeatureProcessor.h
@@ -10,7 +10,6 @@
 
 #include <Atom/Feature/Mesh/MeshFeatureProcessorInterface.h>
 #include <Atom/Feature/Mesh/ModelReloaderSystemInterface.h>
-#include <Atom/Feature/RayTracing/RayTracingFeatureProcessor.h>
 #include <Atom/RHI/TagBitRegistry.h>
 #include <Atom/RPI.Public/Culling.h>
 #include <Atom/RPI.Public/MeshDrawPacket.h>
@@ -21,6 +20,7 @@
 #include <AzCore/Console/Console.h>
 #include <AzFramework/Asset/AssetCatalogBus.h>
 #include <Mesh/MeshInstanceManager.h>
+#include <RayTracing/RayTracingFeatureProcessor.h>
 #include <TransformService/TransformServiceFeatureProcessor.h>
 
 namespace AZ
