+++ Gems/Atom/Feature/Common/Code/Source/RayTracing/RayTracingAccelerationStructurePass.cpp
@@ -6,7 +6,6 @@
  *
  */
 
-#include <Atom/Feature/RayTracing/RayTracingFeatureProcessor.h>
 #include <Atom/RHI/BufferFrameAttachment.h>
 #include <Atom/RHI/BufferScopeAttachment.h>
 #include <Atom/RHI/CommandList.h>
@@ -19,6 +18,7 @@
 #include <Atom/RPI.Public/Scene.h>
 #include <Mesh/MeshFeatureProcessor.h>
 #include <RayTracing/RayTracingAccelerationStructurePass.h>
+#include <RayTracing/RayTracingFeatureProcessor.h>
 
 namespace AZ
 {
