+++ Gems/Atom/Feature/Common/Code/Source/RayTracing/RayTracingPass.cpp
@@ -6,7 +6,6 @@
  *
  */
 
-#include <Atom/Feature/RayTracing/RayTracingFeatureProcessor.h>
 #include <Atom/Feature/RayTracing/RayTracingPass.h>
 #include <Atom/Feature/RayTracing/RayTracingPassData.h>
 #include <Atom/RHI/CommandList.h>
@@ -27,6 +26,7 @@
 #include <Atom/RPI.Reflect/Shader/ShaderAsset.h>
 #include <AzCore/Asset/AssetCommon.h>
 #include <AzCore/Asset/AssetManagerBus.h>
+#include <RayTracing/RayTracingFeatureProcessor.h>
 
 using uint = uint32_t;
 using uint4 = uint[4];
