+++ Gems/Atom/Feature/Common/Code/atom_feature_common_files.cmake
@@ -250,6 +250,8 @@ set(FILES
     Source/PostProcessing/TaaPass.cpp
     Source/PostProcessing/WhiteBalancePass.h
     Source/PostProcessing/WhiteBalancePass.cpp
+    Source/RayTracing/RayTracingFeatureProcessor.h
+    Source/RayTracing/RayTracingFeatureProcessor.cpp
     Source/RayTracing/RayTracingResourceList.h
     Source/RayTracing/RayTracingAccelerationStructurePass.cpp
     Source/RayTracing/RayTracingAccelerationStructurePass.h
